requirements.txt
Django==2.2.19
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
pkg-resources==0.0.0
psycopg2==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
